+++ modules/telegram/src/main/java/io/starter/telegram/cache/MessageCache.java
@@ -1,10 +1,10 @@
-package io.starter.telegram.cash;
+package io.starter.telegram.cache;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 
-import io.starter.telegram.cash.state.MessageState;
+import io.starter.telegram.cache.state.MessageState;
 
 import lombok.Getter;
 import org.springframework.stereotype.Service;
@@ -13,7 +13,7 @@ import org.telegram.telegrambots.meta.api.objects.message.Message;
 
 @Getter
 @Service
-public class MessageCash {
+public class MessageCache {
 
   private final Map<Long, MessageState> botStateMap = new HashMap<>();
