+++ modules/telegram/src/main/java/io/starter/telegram/cache/state/CallbackState.java
@@ -1,4 +1,4 @@
-package io.starter.telegram.cash.state;
+package io.starter.telegram.cache.state;
 
 import lombok.AllArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
