+++ modules/telegram/src/main/java/io/starter/telegram/cache/state/MessageState.java
@@ -1,4 +1,4 @@
-package io.starter.telegram.cash.state;
+package io.starter.telegram.cache.state;
 
 import io.starter.telegram.constants.Emoji;
