+++ modules/telegram/src/main/java/io/starter/telegram/constants/League.java
@@ -1,6 +1,6 @@
 package io.starter.telegram.constants;
 
-import io.starter.telegram.cash.state.CallbackState;
+import io.starter.telegram.cache.state.CallbackState;
 
 public enum League {
