+++ modules/telegram/src/main/java/io/starter/telegram/service/CallbackAnswerService.java
@@ -3,7 +3,7 @@ package io.starter.telegram.service;
 import java.util.List;
 import java.util.Objects;
 
-import io.starter.telegram.cash.state.CallbackState;
+import io.starter.telegram.cache.state.CallbackState;
 import io.starter.telegram.constants.Constants;
 import io.starter.telegram.constants.Emoji;
 import io.starter.telegram.constants.League;
