+++ modules/telegram/src/test/java/io/starter/tests/units/handler/message/BaseMessageTest.java
@@ -1,7 +1,7 @@
 package io.starter.tests.units.handler.message;
 
-import io.starter.telegram.cash.CallbackCash;
-import io.starter.telegram.cash.MessageCash;
+import io.starter.telegram.cache.CallbackCache;
+import io.starter.telegram.cache.MessageCache;
 import io.starter.telegram.dao.UserDao;
 import io.starter.telegram.service.CallbackAnswerService;
 import io.starter.telegram.service.MessageAnswerService;
@@ -26,8 +26,8 @@ public abstract class BaseMessageTest extends BaseTest {
   protected final User user = mock(User.class);
 
   protected final MessageAnswerService messageAnswerService = spy(MessageAnswerService.class);
-  protected final CallbackCash callbackCash = spy(CallbackCash.class);
-  protected final MessageCash messageCash = spy(MessageCash.class);
+  protected final CallbackCache callbackCache = spy(CallbackCache.class);
+  protected final MessageCache messageCache = spy(MessageCache.class);
 
   @BeforeClass
   protected void setUp() {
