+++ modules/telegram/src/test/java/io/starter/tests/units/handler/message/commands/CommandsTest.java
@@ -9,9 +9,9 @@ import org.telegram.telegrambots.meta.api.methods.botapimethods.BotApiMethod;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.testng.annotations.Test;
 
-import static io.starter.telegram.cash.state.MessageState.SETTINGS;
-import static io.starter.telegram.cash.state.MessageState.START;
-import static io.starter.telegram.cash.state.MessageState.WELCOME;
+import static io.starter.telegram.cache.state.MessageState.SETTINGS;
+import static io.starter.telegram.cache.state.MessageState.START;
+import static io.starter.telegram.cache.state.MessageState.WELCOME;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.when;
@@ -21,7 +21,7 @@ public class CommandsTest extends BaseMessageTest {
   @Test(description = "Bot should welcome user when he writes command '/start'")
   void testWhenBotGotCommandStart() {
     UpdateHandler handler = spy(new UpdateHandler(messageAnswerService, callbackAnswerService, userDao));
-    TelegramFacade bot = spy(new TelegramFacade(handler, callbackCash, messageCash));
+    TelegramFacade bot = spy(new TelegramFacade(handler, callbackCache, messageCache));
     String firstName = faker.name().firstName();
 
     when(message.getText()).thenReturn(WELCOME.value);
@@ -38,7 +38,7 @@ public class CommandsTest extends BaseMessageTest {
   @Test(description = "Bot should send message about selecting next command")
   void testWhenBotGotCommandExpandOptions() {
     UpdateHandler handler = spy(new UpdateHandler(messageAnswerService, callbackAnswerService, userDao));
-    TelegramFacade bot = spy(new TelegramFacade(handler, callbackCash, messageCash));
+    TelegramFacade bot = spy(new TelegramFacade(handler, callbackCache, messageCache));
 
     when(message.getText()).thenReturn(START.value);
     BotApiMethod<?> botApiMethod = bot.handleOnUpdate(update);
@@ -53,7 +53,7 @@ public class CommandsTest extends BaseMessageTest {
   @Test(description = "Bot should send message with Settings")
   void testWhenBotGotCommandSettings() {
     UpdateHandler handler = spy(new UpdateHandler(messageAnswerService, callbackAnswerService, userDao));
-    TelegramFacade bot = spy(new TelegramFacade(handler, callbackCash, messageCash));
+    TelegramFacade bot = spy(new TelegramFacade(handler, callbackCache, messageCache));
 
     when(message.getText()).thenReturn(SETTINGS.value);
     BotApiMethod<?> botApiMethod = bot.handleOnUpdate(update);
@@ -67,7 +67,7 @@ public class CommandsTest extends BaseMessageTest {
   @Test(description = "Bot should ignore not supporting messages")
   void testWhenBotGotUnknownCommand() {
     UpdateHandler handler = spy(new UpdateHandler(messageAnswerService, callbackAnswerService, userDao));
-    TelegramFacade bot = spy(new TelegramFacade(handler, callbackCash, messageCash));
+    TelegramFacade bot = spy(new TelegramFacade(handler, callbackCache, messageCache));
     String text = faker.text().text();
 
     when(message.getText()).thenReturn(text);
